Refactor object
