+++ bukkit/src/main/java/com/denizenscript/depenizen/bukkit/support/plugins/LibsDisguisesSupport.java
@@ -4,7 +4,7 @@ import com.denizenscript.depenizen.bukkit.commands.libsdisguises.DisguiseCommand
 import com.denizenscript.depenizen.bukkit.events.libsdisguises.EntityDisguisesScriptEvent;
 import com.denizenscript.depenizen.bukkit.events.libsdisguises.EntityUndisguisesScriptEvent;
 import com.denizenscript.depenizen.bukkit.extensions.libsdisguise.LibsDisguiseEntityExtension;
-import com.denizenscript.depenizen.bukkit.objects.dLibsDisguise;
+import com.denizenscript.depenizen.bukkit.objects.libsdisguises.LibsDisguise;
 import com.denizenscript.depenizen.bukkit.support.Support;
 import net.aufdemrand.denizen.objects.dEntity;
 
@@ -13,7 +13,7 @@ public class LibsDisguisesSupport extends Support {
     public LibsDisguisesSupport() {
         registerScriptEvents(new EntityDisguisesScriptEvent());
         registerScriptEvents(new EntityUndisguisesScriptEvent());
-        registerObjects(dLibsDisguise.class);
+        registerObjects(LibsDisguise.class);
         registerProperty(LibsDisguiseEntityExtension.class, dEntity.class);
         new DisguiseCommands().activate().as("DISGUISE").withOptions("See Documentation.", 1);
     }
